tjBufSizeYUV2(): Fix int overflow w/ large params
The PAD(pw, align) computation was performed using signed integer
arithmetic and could thus overflow if pw + align > INT_MAX.  This was
innocuous, however, because:

- The intermediate value always overflowed to a negative integer.
- When the PAD() macro's bitwise AND operator was applied, the negative
  integer changed to the correct positive integer.

In other words, because PAD() is ultimately a bitwise operation rather
than a mathematical operation, this was purely a UBSan warning rather
than an actual issue.

Fixes #896

+++ turbojpeg.c
@@ -602,11 +602,11 @@ DLLEXPORT unsigned long tjBufSizeYUV2(int width, int align, int height,
   nc = (subsamp == TJSAMP_GRAY ? 1 : 3);
   for (i = 0; i < nc; i++) {
     int pw = tjPlaneWidth(i, width, subsamp);
-    int stride = PAD(pw, align);
+    unsigned long long stride = PAD((unsigned long long)pw, align);
     int ph = tjPlaneHeight(i, height, subsamp);
 
     if (pw < 0 || ph < 0) return -1;
-    else retval += (unsigned long long)stride * ph;
+    else retval += stride * ph;
   }
 #if ULLONG_MAX > ULONG_MAX
   if (retval > (unsigned long long)((unsigned long)-1))
